Added my Web Project Files
When putting my HTML through the validator, 7 error pops up. However all of these errors deal with the same issue of uploading a video to the website. The program said that iframe was invalid, but myself and both of the other 2 TAs looked up other methods to fix the issue and found none. Apparently iframe is not supported by the specific HTML version, however the TAs said this should not be a major issue and that I should be fine turning it in. However if you want me to change it, I can. Thank you very much.